Implement exit_codes in FLEUR calculation and parser
Note that 'successfull' switch of the parser is depricated
thus the parser behaviour is changed. It used to try to
parse as much as possible even if calculation was failed.
Now it instantly throws a corresponding exit_code if some
important part of parsing procedure (like out.xml is missing)
is failed.