Implement exit_codes into scf workchain
Slight changes are done in the spec.outline since exit_codes do
not work properly inside if_ or while_ statements. They do not
affect the logic the workchain but introduce some level of complexity
in understanding.
Similarly to fleur calculation, 'successfull' output parameter is
depricated because now an exit_code indicates if a workchain
was successful.