Add support of relax.xml and clean up in fleur calc
FLEUR calculation and parser supports now relax.xml as a part
of fleurinp data type. The main changes are in the parser:
now it checks if relax.xml was retrieved and changed; then
parses it. Results of parsing are stored in the output node
called 'relax_parameters'. If there is no relax.xml in input
nor retrieved folder the calculation behaves the same.
Moreover, the code was cleaned up. Now it is closer to PEP8
standarts.