Initial commit for refactored relax wc
New relax wc relies on the new versions of scf wc and
fleur calculation. It has 4 input modes:
1. fluerinp -> take inp.xml from fleurinp
2. fleurinp + remote -> take inp.xml from fleurinp
   and chd1 from remote
3. remote -> take inp.xml (and relax.xml if exists)
   and cdn1 from remote
4. structure -> run inpgen
These modes have priority accorting to it's number in
the list above.
The workchain submits scf wc in the 'force' mode several
times until the largest forse is not small enogh.
Generates an optimized structure as an output.