Split mae wc into two workchains: force and conv
Mae wc was split because the force theorem and conv
calculations differs significantily in terms of
input and output.
The force workchain uses the same pattern as other
force theorem workchian SSDisp and DMI. They will
be probably rewritten to inherit from the base
force theorem workchian.
The conv workchain is not implemented yet.