Add set_species_label and change_atomgr_att_label xml utils
It is usually the case when a user wants to change some parameters
of a particular atom and it is not known which specie or
atomGroup corresponds to the atom.
Two new methods, change_atomgr_att_label and set_species_label
allow one to change parameters knowing label only.
To use this methods correctly, it is important to set a proper
name of an atom that one wants to change. For intance,
name='Fe241' creates an iron atom with a label '241' and
make this atom to belong to the separate specie and atomGroup.
Passing '241' as an atom label, it is possible to change
required parameter in specie or atomGroup.