Implement smarter analysis of memory issues
Now every fleur submission executes
$ cat /proc/meminfo > mem_avail.txt
and the mem_avail.txt file is copied to the local
machine and parsed. We also parse amount of memory
which was used by a MPI process that caused a crash.
If a rough estimation of total used memory is
bigger than 93% of available memory, I consider FLEUR
to be crashed due to memory issues and resubmit it with
twice larger amount of nodes.