Correct errors found by Pylint, reduce six methods
Pre-commit is running through now, except for pylint.
Errors found by pylint are corrected. For now ignore most common
warnings and fix the others and important ones.
Uncommented ones that are todo are marked with fixme in the pylintrc file.
In this process get rid of some six methods, since python2 is not
supported anylonger by aiida-fleur.