:bug: Make fleurinpData handle files from nodes
The last commit introduced a bug, which caused
fleurinpData add_path to fail if the file is written from a given node.
To solve this and still keep the context manager, add_path was
refactored. The inp.xml file is now multiple times read completely and
in the case of a node it is read as a byte string and written again.