Make Aiida-testing dependency optional
For a release we currently cannot depend on
aiida-testing, but we need it for the tests only.
So we do not have it as a dependency as long as there is no release.
Instead we have the option to switch on and of regression tests.
if aiida-testing is not installed we skip these tests.