Add kpoints_distance to scf wf_para
this key enforces a minimum kpoint distance, which will generate a
mesh using and will merge this into optional given calc_parameters for inpgen
Therefore the hope is that one can control this way an structure independed
convergence parameter. Which is useful for protocols or simple error handling in
non convergence cases, i.e just make the mesh denser. This relates to issue #26.