Introduce cli launch command
First implementation of launch command for various workchains.
Has still to be made robust. This commit also includes fixes by precommit,
a dict util helper to remove keys with None values from dicts.
For the CLI options are introduced, also one option representing a structure,
which can be given as a node a are file of some format parsable by ase.