Persist Kind/Species information through relaxation wc (#102)
In the relaxation case also parse information on the atomtypes and
the species names. The relaxation wc presists these species names.
This is important when one relaxes magnetic systems or any structure,
which had several kinds at the beginning. Relaxed to issue # 101.

* Add fleur parser test for relax

* Make CI stay on ubuntu 18.04 for now, since 20.04 causes problems

Verifying the rabbitmq install.
We will switch back to ubuntu-latest if github is finished with its migration