Port fleurinp to MaXR5, fix get_kpoints
The kpoint paths and list forms of MaXR5 have changed.
For now I adjusted the path. Fleurinpdata might be back compatible.
Fixed pretty print of fleurinpmodifier show, since to string returns a
bit string, which was not converted for python3. Also paths for dos
switches. This is not backcompatible, without xpaths from schema files.