Add `python_requires` and relax `numpy` requirements
Add the `python_requires` key to the `setup.json` which will inform
`pip` what minimum version of Python is supported. The upper requirement
of `numpy` is removed, since `pip` should automatically prevent
installing a version of `numpy` that requires a higher version of Python
than is installed.

Finally, the requirement for `aiida-core` is simplified by using the
compatibility operator `aiida-core~=1.3` which means exactly the same as
`aiida-core>=1.3,<2.0` but is more succinct and readable.