Improve error messages for Fleur_inputgenParser
The parser will now log the parser_info output from the inpxml_parser in masci-tools. This is done by initializing the fleurinpdata object first and then adding the inp.xml. Additionally the parser_info extra is now set in a finally clause around the inpxml_parser to make sure that it's set