Adjust default for `optimize_resources` in options for `FleurBaseWorkChain` (#129)
The default values were different between the `FleurSCFWorkChain (True)` and `FleurBaseWorkChain (False)`.
This led to the behaviour that submitting the `FleurBaseWorkChain` not through scf would unexpectedly not
adjust the number of MPI/OMP to the number of kpoints. This step should be the default behaviour and should
only be switched off if there is some problem with it.