Remove unused arguments in plot_fleur_sn/mn, Added banddos workchain visualization, moved orbcontrol vis to common routine
Setting the default values in multiple places can lead to confusion. E.g. the backend was set in the main function
to matplotlib by default. All other functions use bokeh as default. Ideally the backend is not specified until we reach masci-tools.
However this is not yet possible for all functions