Improve time limit error handler in FLeurBaseWorkChain.
Introduced analogous ERROR_TIME_LIMIT_NO_SOLUTION exit code to  not enough memory case to represent
the case, where the calculation should not be restarted.
Previously the workchain would just exit with no indication that something is wrong

Also if the calculation is restarted using the remote folder from the previous calculation
a check is added if there is a charge density file to be uploaded. Otherwise the workchain will
except in the transport stage, which might be confusing for users since the time limit error is not
immediately apparent