Improve out of memory/walltime error handlers
The out of memory error can also occur after several iterations if the
mixing_history no longer fits into memory. For this reason atleast check
for a `cdn_last.hdf` file and if found reuse it

The out of walltime memory handler no longer instantly kills the
workflow if no charge density is available. The charge density is
ignored in this case